overhaul: config via JSON not env-var, and lots more

+++ src/sshsvc/hup_sshd.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
-SSHDPID=$(cat /var/run/sshd.pid)
-kill -HUP $SSHDPID
-echo "hup_sshd.pid: just HUP'd sshd ($PID)" >&2
+if SSHDPID=$(cat /var/run/sshd.pid 2>/dev/null); then
+	kill -HUP $SSHDPID
+	echo "hup_sshd.pid: just HUP'd sshd ($PID)" >&2
+fi
